- added constructor to image for specifying a filename - started with some simple speed tests of the convolution against opencv functions
A    visionTools/tests/speedTests
A    visionTools/tests/speedTests/speedTests.cpp
A    visionTools/tests/speedTests/ImageProcessing.h
A    visionTools/tests/speedTests/CMakeLists.txt
M    visionTools/tests/CMakeLists.txt
M    visionTools/cvt/gfx/Image.cpp
M    visionTools/cvt/gfx/Image.h
M    visionTools/cvt/io/ImageIO.cpp
M    visionTools/cvt/vision/ESM.h


git-svn-id: https://svnknoll.informatik.tu-muenchen.de/kloses/visionTools@914 f7bcf5df-4c73-0410-a0d8-b61c104701fe

+++ cvt/vision/ESM.h
@@ -220,7 +220,7 @@ namespace cvt {
 		T weight;
 		int height = _temp.height();		
 		while( height-- ){
-			for( int x = 0; x < _temp.width(); x++ ){
+			for( size_t x = 0; x < _temp.width(); x++ ){
 				_costs+= costFunc.cost( w[ x ] - t[ x ], weight );
 			}
 			w += wstride;
